bot/requirements.txt
setuptools==78.1.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
redis==3.5.3
